repos_skeleton/alpha-crucible-data-reddit/main.py
#!/usr/bin/env python3
"""
Alpha Crucible Data - Reddit
Fetches Reddit data and stores it in ORE database.
"""

import os
import sys
import logging
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


def main():
    """Main execution function."""
    try:
        logger.info("Starting Reddit data fetch...")
        
        # TODO: Implement Reddit API data fetching
        # Example structure:
        # 1. Connect to Reddit API
        # 2. Fetch data (posts, comments, etc.)
        # 3. Transform data
        # 4. Connect to ORE database
        # 5. Insert data into ORE database
        
        # Placeholder: Print hello world for now
        logger.info("Hello World from alpha-crucible-data-reddit!")
        logger.info("This is a skeleton implementation.")
        logger.info("Replace this with actual Reddit data fetching logic.")
        
        # Example database connection (uncomment when ready)
        # ore_db_url = os.getenv('ORE_DATABASE_URL')
        # if not ore_db_url:
        #     raise ValueError("ORE_DATABASE_URL not set in environment")
        # 
        # # Connect and insert data
        # # ... your implementation here ...
        
        logger.info("Reddit data fetch completed successfully")
        return 0
        
    except Exception as e:
        logger.error(f"Error during execution: {e}", exc_info=True)
        return 1


if __name__ == "__main__":
    exit_code = main()
    sys.exit(exit_code)

