repos_skeleton/alpha-crucible-data-reddit/requirements.txt
# Database
psycopg2-binary>=2.9.9,<3.0.0

# Data Processing
pandas>=2.1.4,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Reddit API (example - adjust based on your needs)
# praw>=7.7.0,<8.0.0

# Logging
python-json-logger>=2.0.7,<3.0.0

